# refactor: Make it thread-safe