--- a/src/pdf417/decoder/detection_result_column.rs
+++ b/src/pdf417/decoder/detection_result_column.rs
@@ -14,17 +14,17 @@
  * limitations under the License.
  */
 
-use std::{fmt::Display, rc::Rc};
+use std::{fmt::Display, sync::Arc};
 
 use super::{BoundingBox, Codeword, DetectionRXingResultRowIndicatorColumn};
 
 const MAX_NEARBY_DISTANCE: u32 = 5;
 
 pub trait DetectionRXingResultColumnTrait {
-    fn new_column(boundingBox: Rc<BoundingBox>) -> DetectionRXingResultColumn
+    fn new_column(boundingBox: Arc<BoundingBox>) -> DetectionRXingResultColumn
     where
         Self: Sized;
-    fn new_with_is_left(boundingBox: Rc<BoundingBox>, isLeft: bool) -> DetectionRXingResultColumn
+    fn new_with_is_left(boundingBox: Arc<BoundingBox>, isLeft: bool) -> DetectionRXingResultColumn
     where
         Self: Sized;
     fn getCodewordNearby(&self, imageRow: u32) -> &Option<Codeword>;
@@ -48,7 +48,7 @@ pub struct DetectionRXingResultColumn {
 }
 
 impl DetectionRXingResultColumnTrait for DetectionRXingResultColumn {
-    fn new_column(boundingBox: Rc<BoundingBox>) -> DetectionRXingResultColumn {
+    fn new_column(boundingBox: Arc<BoundingBox>) -> DetectionRXingResultColumn {
         DetectionRXingResultColumn {
             boundingBox: BoundingBox::from_other(boundingBox.clone()),
             codewords: vec![None; (boundingBox.getMaxY() - boundingBox.getMinY() + 1) as usize],
@@ -56,7 +56,7 @@ impl DetectionRXingResultColumnTrait for DetectionRXingResultColumn {
         }
     }
 
-    fn new_with_is_left(boundingBox: Rc<BoundingBox>, isLeft: bool) -> DetectionRXingResultColumn {
+    fn new_with_is_left(boundingBox: Arc<BoundingBox>, isLeft: bool) -> DetectionRXingResultColumn {
         DetectionRXingResultColumn {
             boundingBox: BoundingBox::from_other(boundingBox.clone()),
             codewords: vec![None; (boundingBox.getMaxY() - boundingBox.getMinY() + 1) as usize],