--- a/src/qrcode/decoder/qrcode_decoder.rs
+++ b/src/qrcode/decoder/qrcode_decoder.rs
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-use std::{collections::HashMap, rc::Rc};
+use std::{collections::HashMap, sync::Arc};
 
 /**
  * <p>The main class which implements QR Code decoding -- as opposed to locating and extracting
@@ -117,7 +117,7 @@ pub fn decode_bitmatrix_with_hints(
         let mut result = decode_bitmatrix_parser_with_hints(&mut parser, hints)?;
 
         // Success! Notify the caller that the code was mirrored.
-        result.setOther(Some(Rc::new(QRCodeDecoderMetaData::new(true))));
+        result.setOther(Some(Arc::new(QRCodeDecoderMetaData::new(true))));
 
         Ok(result)
     };