… src/datamatrix/encoder/high_level_encode_test_case.rs
… src/pdf417/decoder/ec/modulus_poly.rs
… src/qrcode/encoder/minimal_encoder.rs
… tests/common/multiimage_span.rs