The current implementation of rxing uses Rc for reference counting, which is not thread-safe. This limitation restricts the library’s use in multi-threaded applications where safe concurrent access to shared resources is required.

Passed all test cases successfully.