@hschimke 2024-07-30
Hello, I'm going to admit my lack of familiarity with thread-safety in rust. Does the entire library need to be fully thread-safe in order to be used in a multi-threaded context? I understand why you might want the things passed in and out of the library to be fully thread-safe, but I'm unclear on if every internal bit has to be Arc. I'm curious about the context where this doesn't work.