@hschimke 2024-07-30
Actually now that I have read a bit more and tested it I see the point. I'll review this later this week. Thanks