@Iceapinan 2024-07-30
Correct. The library must be thread-safe to be used in a multi-threaded context. I recommend using Arc<T> instead of Rc<T> as it provides library users with more flexibility.