@hschimke 2024-07-30
I'm about half done with the review. Do you have benchmarks for the before-after change (I'd like to include them in the patch notes).